+++ api/v1/mdb/mongodbbuilder.go
@@ -177,9 +177,8 @@ func (b *MongoDBBuilder) Build() *MongoDB {
 
 func defaultMongoDB(resourceType ResourceType) *MongoDBBuilder {
 	spec := MongoDbSpec{
-		Version:        "4.0.0",
-		ResourceType:   resourceType,
-		ConnectionSpec: ConnectionSpec{Project: "test-project"},
+		Version:      "4.0.0",
+		ResourceType: resourceType,
 	}
 	mdb := &MongoDB{Spec: spec, ObjectMeta: metav1.ObjectMeta{Name: "test-mdb", Namespace: "testNS"}}
 	mdb.InitDefaults()